requirements.txt
anthropic>=0.25.0
requests>=2.31.0
html2text>=2020.1.16
python-dotenv>=1.0.0
watchdog>=4.0.0
